split up cpp and removed unused resources
end screen updated as well
