--- resource/dungeon_narrative.txt
@@ -1,11 +1,0 @@
-You wake in a moonlit cell.
-
-Motes of dust float through the air, flickering in and out of existence
-as they cross beneath the shadows of heavy iron bars.
-
-Through a narrow window set high in the crumbling stone wall you can glimpse
-a courtyard and within it stands a great oak tree.
-
-Your choices are as vast as its sprawling branches. 
-
-Where will your journey lead you?
