--- resource/dungeon_options.txt
@@ -1,1 +1,0 @@
-I am the dungeon option folder, there is nothing in me.
