--- resource/tavern_narrative.txt
@@ -1,1 +1,0 @@
-I am the tavern narrative file!
