--- resource/tavern_options.txt
@@ -1,1 +1,0 @@
-I am the tavern option file!
